refactor : MentorDetialContent 뷰로직 분리했습니다
